feat: faqの処理のフック化、エディターから不要な項目を削除。
